app/requirements.txt
anyascii==0.3.2
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.2
autobahn==23.1.2
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
cloudinary==1.30.0
colorama==0.4.6
convertdate==2.4.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.3
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-rest-swagger==2.2.0
django-taggit==3.1.0
django-templated-mail==1.1.1
django-timezone-field==5.0
django-treebeard==4.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
draftjs-exporter==2.1.7
drf-yasg==1.21.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
feedgen==0.9.0
filelock==3.9.0
hijri-converter==2.2.4
holidays==0.21.13
html5lib==1.1
icalendar==5.0.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
korean-lunar-calendar==0.3.1
l18n==2021.3
ls.joyous==1.4.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
num2words==0.5.12
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==21.3
Pillow==9.3.0
pipenv==2022.12.19
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
PyMeeus==0.5.12
pyOpenSSL==23.1.1
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.4
sqlparse==0.4.3
telepath==0.3
tomli==2.0.1
Twisted==22.10.0
# twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
Willow==1.4.1
zipp==3.10.0
zope.interface==6.0
